test: extract and add stories for in chat download-buttons
