+++ apps/nextjs/src/components/AppComponents/Chat/chat-list/demo-limit-message.stories.tsx
@@ -14,8 +14,6 @@ const meta: Meta<typeof DemoLimitMessage> = {
         <Story />
       </ChatModerationProvider>
     ),
-    // ChatDecorator,
-    // DemoDecorator,
   ],
 };
